[feature/#60] api connect

+++ client/app/src/main/java/com/example/circuler/presentation/ui/history/navigation/HistoryNavigation.kt
@@ -16,7 +16,7 @@ fun NavController.navigateToHistory(navOptions: NavOptions) {
 }
 
 fun NavGraphBuilder.historyNavGraph(
-    navigateToSubmit: () -> Unit
+    navigateToSubmit: (Int) -> Unit
 ) {
     composable<MainTabRoute.RequestPackage> {
         HistoryRoute(
